pyproject.toml
[project]
name = "travelagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ddgs>=9.5.0",
    "duckduckgo-search>=8.1.1",
    "folium>=0.20.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.2",
    "openrouteservice>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]
